--- envosaurus.go
@@ -1,9 +1,0 @@
-package main
-
-import (
-	"github.com/tentwentyfive/envosaurus/cmd"
-)
-
-func main() {
-	cmd.Execute()
-}
